Refactor sending emails
